Implementation of Unclassified Datasets' Features and Manipulation w/ JS & PHP: 1) pretrainedModels - Frontend V2: Uploading, Selection, Downloading, Deletion and Preview of Unclassified Datasets. 2) pretrainedModels - Backend V2: 5 New Endpoints (receiveUnclassifiedDatasets.php [GET], uploadUnclassifiedDataset.php [POST], multilabelUnclassifiedDatasetContent.php [GET], destroyUnclassifiedDataset.php [DELETE], downloadUnclassifiedDataset.php [GET]).

+++ server/php/api/destroyModel.php
@@ -37,16 +37,30 @@
 
     // $file_path = "../../py/users/$hash_user/models/$file"; # << CORRECT (for later addition)
     $file_path = "../../py/users/$file"; # << ΕΓΩ ΤΟ ΈΒΑΛΑ 
-
     if(!file_exists($file_path)) {
         header("HTTP/1.1 400 Bad Request");
         print json_encode(['errormesg'=>"File doesn't exist."]);
         exit;
     }
     
+    $trf_file = str_replace('.pkl', '', $file);
+    // $transformation_path = "../../py/users/transformations/$hash_user/" . $trf_file . "_transformation.pkl"; # << CORRECT (for later addition)
+    $transformation_path = "../../py/users/transformations/" . $trf_file . "_transformation.pkl";
+    if(!file_exists($transformation_path)) {
+        header("HTTP/1.1 400 Bad Request");
+        print json_encode(['errormesg'=>"File doesn't exist."]);
+        exit;
+    }
+
     $delete = unlink($file_path);
-    
     if(!$delete) {
+        header("HTTP/1.1 400 Bad Request");
+        print json_encode(['errormesg'=>"Unable to delete this file."]);
+        exit;
+    }
+
+    $delete2 = unlink($transformation_path);
+    if(!$delete2) {
         header("HTTP/1.1 400 Bad Request");
         print json_encode(['errormesg'=>"Unable to delete this file."]);
         exit;
